requirements.txt
absl-py==2.1.0
attrs==23.2.0
cffi==1.16.0
contourpy==1.1.1
cvzone==1.6.1
cycler==0.12.1
flatbuffers==24.3.25
fonttools==4.53.0
importlib-metadata==7.1.0
importlib-resources==6.4.0
jax==0.4.13
kiwisolver==1.4.5
matplotlib==3.7.5
mediapipe==0.10.11
ml-dtypes==0.2.0
numpy==1.24.4
opencv-contrib-python==4.10.0.82
opencv-python==4.10.0.82
opt-einsum==3.3.0
packaging==24.1
pillow==10.3.0
protobuf==3.20.3
pycparser==2.22
pynput==1.7.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.10.1
six==1.16.0
sounddevice==0.4.7
zipp==3.19.2
